Move IEditor, IDisplay and IInputHistory back to core
The Eclipse implementation turned out to be similar, so the interfaces are
useful to have in core.
