+++ org.metaborg.spoofax.shell.console/src/main/java/org/metaborg/spoofax/shell/client/console/impl/ConsoleReplModule.java
@@ -4,19 +4,19 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 
-import org.metaborg.spoofax.shell.client.console.IDisplay;
-import org.metaborg.spoofax.shell.client.console.IEditor;
-import org.metaborg.spoofax.shell.client.console.IInputHistory;
+import org.metaborg.spoofax.shell.client.IDisplay;
+import org.metaborg.spoofax.shell.client.IEditor;
+import org.metaborg.spoofax.shell.client.IInputHistory;
 import org.metaborg.spoofax.shell.client.console.commands.ExitCommand;
 import org.metaborg.spoofax.shell.client.console.impl.history.JLine2InputHistory;
 import org.metaborg.spoofax.shell.client.console.impl.history.JLine2PersistentInputHistory;
 import org.metaborg.spoofax.shell.client.console.impl.hooks.ConsoleMessageHook;
 import org.metaborg.spoofax.shell.client.console.impl.hooks.ConsoleResultHook;
 import org.metaborg.spoofax.shell.commands.IReplCommand;
+import org.metaborg.spoofax.shell.client.hooks.IMessageHook;
+import org.metaborg.spoofax.shell.client.hooks.IResultHook;
 import org.metaborg.spoofax.shell.core.IRepl;
 import org.metaborg.spoofax.shell.core.ReplModule;
-import org.metaborg.spoofax.shell.hooks.IMessageHook;
-import org.metaborg.spoofax.shell.hooks.IResultHook;
 
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
