+++ org.metaborg.spoofax.shell.console/src/test/java/org/metaborg/spoofax/shell/client/console/ConsoleReplTest.java
@@ -19,6 +19,7 @@ import java.io.UnsupportedEncodingException;
 
 import org.junit.Test;
 import org.metaborg.core.MetaborgException;
+import org.metaborg.spoofax.shell.client.IEditor;
 import org.metaborg.spoofax.shell.client.console.commands.ExitCommand;
 import org.metaborg.spoofax.shell.client.console.impl.ConsoleRepl;
 import org.metaborg.spoofax.shell.client.console.impl.ConsoleReplModule;
