+++ org.metaborg.spoofax.shell.console/src/test/java/org/metaborg/spoofax/shell/client/console/MockModule.java
@@ -1,5 +1,7 @@
 package org.metaborg.spoofax.shell.client.console;
 
+import org.metaborg.spoofax.shell.client.IDisplay;
+import org.metaborg.spoofax.shell.client.IEditor;
 import org.metaborg.spoofax.shell.invoker.ICommandInvoker;
 
 import com.google.inject.AbstractModule;
