+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/client/hooks/IResultHook.java
@@ -1,4 +1,4 @@
-package org.metaborg.spoofax.shell.hooks;
+package org.metaborg.spoofax.shell.client.hooks;
 
 import java.util.function.Consumer;
