+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/client/hooks/package-info.java
@@ -2,4 +2,4 @@
  * The {@link org.metaborg.spoofax.shell.core.IRepl} interface uses hooks to pass results back to
  * the client. This package defines two such hooks.
  */
-package org.metaborg.spoofax.shell.hooks;
+package org.metaborg.spoofax.shell.client.hooks;
