+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/client/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * The client package contains interfaces that are supposed to be implemented by clients.
+ */
+package org.metaborg.spoofax.shell.client;
